ft_strlcpy OK

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: tkitahar <tkitahar@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/21 18:06:45 by tkitahar          #+#    #+#             */
-/*   Updated: 2024/04/22 15:31:05 by tkitahar         ###   ########.fr       */
+/*   Updated: 2024/04/23 18:58:26 by tkitahar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (*s++)
